requirements.txt
langchain_experimental==0.0.64
langchain-huggingface==0.0.3
langchain~=0.2.12
langchain-weaviate
python-dotenv==1.0.1
pydantic==2.8.2
sentence-transformers==3.0.1
tiktoken
loguru~=0.7.2
fastapi==0.104.*
python-multipart
protobuf==3.20.3
numpy==1.25.0
torchvision==0.19.0
uvicorn~=0.21.1
starlette~=0.27.0
requests~=2.32.3